refactor(form/pta): Fix linter problems
